fix: Add text resources to create code list dialog
