+++ frontend/libs/studio-content-library/src/ContentLibrary/LibraryBody/pages/CodeListPage/CodeLists/EditCodeList/EditCodeList.tsx
@@ -1,4 +1,4 @@
-import type { CodeList, CodeListEditorTexts } from '@studio/components';
+import type { CodeList, CodeListEditorTexts, TextResource } from '@studio/components';
 import {
   StudioDeleteButton,
   StudioModal,
@@ -6,7 +6,7 @@ import {
   StudioCodeListEditor,
   StudioToggleableTextfield,
 } from '@studio/components';
-import React, { useCallback, useMemo } from 'react';
+import React from 'react';
 import { useTranslation } from 'react-i18next';
 import type { CodeListWithMetadata } from '../../CodeListPage';
 import { useCodeListEditorTexts } from '../../hooks/useCodeListEditorTexts';
@@ -16,42 +16,31 @@ import { useInputCodeListNameErrorMessage } from '../../hooks/useInputCodeListNa
 import classes from './EditCodeList.module.css';
 import type { CodeListIdSource } from '../../types/CodeListReference';
 import { CodeListUsages } from './CodeListUsages/CodeListUsages';
-import type { TextResources } from '../../../../../../types/TextResources';
-import type { TextResource } from '../../../../../../types/TextResource';
-import type { TextResourceWithLanguage } from '../../../../../../types/TextResourceWithLanguage';
-import { createTextResourceWithLanguage, getTextResourcesForLanguage } from './utils';
 
 export type EditCodeListProps = {
   codeList: CodeList;
   codeListTitle: string;
+  onChangeTextResource?: (textResource: TextResource) => void;
   onDeleteCodeList: (codeListId: string) => void;
   onUpdateCodeListId: (codeListId: string, newCodeListId: string) => void;
   onUpdateCodeList: (updatedCodeList: CodeListWithMetadata) => void;
-  onUpdateTextResource?: (textResource: TextResourceWithLanguage) => void;
   codeListNames: string[];
   codeListSources: CodeListIdSource[];
-  textResources?: TextResources;
+  textResources?: TextResource[];
 };
-
-const language: string = 'nb'; // Todo: Let the user choose the language: https://github.com/Altinn/altinn-studio/issues/14572
 
 export function EditCodeList({
   codeList,
   codeListTitle,
+  onChangeTextResource,
   onDeleteCodeList,
   onUpdateCodeListId,
   onUpdateCodeList,
-  onUpdateTextResource,
   codeListNames,
   codeListSources,
   textResources,
 }: EditCodeListProps): React.ReactElement {
   const editorTexts: CodeListEditorTexts = useCodeListEditorTexts();
-
-  const textResourcesForLanguage = useMemo(
-    () => getTextResourcesForLanguage(language, textResources),
-    [textResources],
-  );
 
   const handleCodeListChange = (updatedCodeList: CodeList): void => {
     const updatedCodeListWithMetadata = updateCodeListWithMetadata(
@@ -62,14 +51,6 @@ export function EditCodeList({
   };
 
   const handleDeleteCodeList = (): void => onDeleteCodeList(codeListTitle);
-
-  const handleChangeTextResource = useCallback(
-    (textResource: TextResource) => {
-      const updatedTextResource = createTextResourceWithLanguage(language, textResource);
-      onUpdateTextResource?.(updatedTextResource);
-    },
-    [onUpdateTextResource],
-  );
 
   const codeListHasUsages = codeListSources.length > 0;
   const isCodeListEditable = codeListSources.length === 0;
@@ -86,9 +67,9 @@ export function EditCodeList({
         codeList={codeList}
         onAddOrDeleteItem={handleCodeListChange}
         onBlurAny={handleCodeListChange}
-        onChangeTextResource={handleChangeTextResource}
+        onChangeTextResource={onChangeTextResource}
         texts={editorTexts}
-        textResources={textResourcesForLanguage}
+        textResources={textResources}
       />
       <CodeListButtons
         codeListHasUsages={codeListHasUsages}
