+++ frontend/libs/studio-content-library/src/index.ts
@@ -5,3 +5,6 @@ export type {
   CodeListIdSource,
   CodeListReference,
 } from './ContentLibrary/LibraryBody/pages';
+export type { TextResource } from './types/TextResource';
+export type { TextResourceWithLanguage } from './types/TextResourceWithLanguage';
+export type { TextResources } from './types/TextResources';
